+++ core/src/main/java/net/momirealms/craftengine/core/pack/host/ResourcePackHost.java
@@ -1,5 +1,6 @@
 package net.momirealms.craftengine.core.pack.host;
 
+import java.nio.file.Path;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 
@@ -8,4 +9,6 @@ public interface ResourcePackHost {
     CompletableFuture<ResourcePackDownloadData> requestResourcePackDownloadLink(UUID player);
 
     ResourcePackDownloadData getResourcePackDownloadLink(UUID player);
+
+    CompletableFuture<Boolean> upload(Path resourcePackPath);
 }
