+++ core/src/main/java/net/momirealms/craftengine/core/pack/host/impl/SelfHost.java
@@ -3,6 +3,7 @@ package net.momirealms.craftengine.core.pack.host.impl;
 import net.momirealms.craftengine.core.pack.host.ResourcePackDownloadData;
 import net.momirealms.craftengine.core.pack.host.ResourcePackHost;
 
+import java.nio.file.Path;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 
@@ -15,6 +16,11 @@ public class SelfHost implements ResourcePackHost {
 
     @Override
     public ResourcePackDownloadData getResourcePackDownloadLink(UUID player) {
+        return null;
+    }
+
+    @Override
+    public CompletableFuture<Boolean> upload(Path resourcePackPath) {
         return null;
     }
 }
